Moved TF IDF search from postgres to EF query in C#

+++ Domain/Models/Posts.cs
@@ -3,7 +3,7 @@
     public class Posts
     {
         public int Id { get; set; }
-        public decimal Rank { get; set; }
+        public double Rank { get; set; }
         public string Title { get; set; }
         public string Body { get; set; }
         public int ParentId { get; set; }
