+++ WebService/DTOs/PostsSearchListDto.cs
@@ -3,7 +3,7 @@
 public class PostsSearchListDto
 {
     public string ThreadLink { get; set; }
-    public decimal Rank { get; set; }
+    public double Rank { get; set; }
     public int PostId { get; set; }
     public string QuestionTitle { get; set; }
     public string PostBody { get; set; }
